prepare for installer
