+++ utilities/resource_path.py
@@ -1,5 +1,6 @@
 """This module contains the ResourcePath class."""
 import os
+import sys
 from utilities import constants
 
 
@@ -16,3 +17,13 @@ class ResourcePath:
         if not os.path.exists(app_data_path):
             os.makedirs(app_data_path)
         return os.path.join(os.getenv('PROGRAMDATA'), relative_path)
+
+    def resource_path(self, relative_path):
+        """ Get absolute path to resource, works for dev and for PyInstaller """
+        try:
+            base_path = getattr(sys, '_MEIPASS',
+                                os.path.dirname(os.path.abspath(__file__)))
+        except FileNotFoundError:
+            base_path = os.path.abspath(".")
+
+        return os.path.join(base_path, relative_path)
