+++ window_config/window_config.py
@@ -1,5 +1,6 @@
 """This module contains the class to configure the main window of the application."""
 import utilities.constants as constants
+from utilities.resource_path import ResourcePath
 
 
 class WinConfig:
@@ -15,7 +16,8 @@ class WinConfig:
             f"{constants.WINDOW_WIDTH}x{constants.WINDOW_HEIGHT}")
         self.root.configure(background=constants.WINDOW_BG_CLR)
         self.root.resizable(constants.WINDOW_HEIGHT, constants.WIDTH_RESIZABLE)
-        self.root.iconbitmap(constants.ASSET_ICON)
+        self.root.iconbitmap(
+            ResourcePath.resource_path(self, constants.ASSET_ICON))
 
         self.top_frame.config(background=constants.BAR_BG_CLR)
         self.top_frame.place(relwidth=0.97,
